+++ www/go/core/views/extjs3/links/LinksDetailPanel.js
@@ -9,10 +9,10 @@ go.links.DetailPanel = Ext.extend(Ext.Panel, {
 			baseParams: {
 				limit: this.limit,
 				position: 0,
-				calculateTotal:true,
-				filter: {
-					entities: [{name: this.link.entity, filter: this.link.filter}]
-				}
+				calculateTotal:true,				
+			},
+			filters: {
+				toEntity: {entities: [{name: this.link.entity, filter: this.link.filter}]}
 			},
 			fields: [
 				'id', 
@@ -35,7 +35,7 @@ go.links.DetailPanel = Ext.extend(Ext.Panel, {
 		var tpl = new Ext.XTemplate('<div class="icons"><tpl for=".">\
 				<p data-id="{id}">\
 				<tpl if="xindex === 1">\
-					<i class="label ' + this.link.iconCls + ' hide-on-hover" ext:qtip="{toEntity}"></i>\
+					<i class="label ' + this.link.iconCls + '" ext:qtip="{toEntity}"></i>\
 				</tpl>\
 				<tpl for="to">\
 				<a>{name}</a>\
@@ -139,8 +139,11 @@ go.links.DetailPanel = Ext.extend(Ext.Panel, {
 		
 		this.hide();
 		
-		this.store.baseParams.filter.entity = dv.entity ? dv.entity : dv.entityStore.entity.name, //dv.entity exists on old DetailView or display panels
-		this.store.baseParams.filter.entityId = dv.model_id ? dv.model_id : dv.currentId; //model_id is from old display panel
+		this.store.setFilter('fromEntity', {
+			entity: dv.entity ? dv.entity : dv.entityStore.entity.name, //dv.entity exists on old DetailView or display panels
+			entityId: dv.model_id ? dv.model_id : dv.currentId //model_id is from old display panel
+		});
+
 		this.store.baseParams.position = 0;
 		this.store.load();
 	}
@@ -150,10 +153,18 @@ go.links.getDetailPanels = function() {
 	
 	var panels = [];
 	
-	go.Entities.getLinkConfigs().forEach(function (e) {		
-		panels.push(new go.links.DetailPanel({
-			link: e
-		}));
+	go.Entities.getLinkConfigs().forEach(function (e) {
+		if(e.linkDetailCards) {
+			var clss = e.linkDetailCards();
+			if(!Ext.isArray(clss)) {
+				clss = [clss];
+			}
+			panels = panels.concat(clss);
+		} else {
+			panels.push(new go.links.DetailPanel({
+				link: e
+			}));
+		}
 	});
 	
 	return panels;