+++ www/modules/calendar/Calendar.js
@@ -2189,13 +2189,41 @@ go.Modules.register("legacy", 'calendar', {
 	iconCls : 'go-tab-icon-calendar',
 	entities: [{
 			name: "Event",
+			
 			links: [{
+				iconCls: 'entity Event orange',
+
 				linkWindow: function() {
 					return GO.calendar.showEventDialog();
 				},
 				linkDetail: function() {
 					return new GO.calendar.EventPanel();
-				}					
+				},
+				linkDetailCards: function() {
+					var forth = new go.links.DetailPanel({
+						link: {
+							title: t("Forthcoming events"),
+							iconCls: 'icon ic-event orange',
+							entity: "Event",
+							filter: null
+						}
+					});
+
+					forth.store.setFilter('forthcomming', {forthComingEvents: true});
+
+					var past = new go.links.DetailPanel({						
+						link: {
+							title: t("Past events"),
+							iconCls: 'icon ic-event orange',
+							entity: "Event",
+							filter: null
+						}
+					});
+
+					past.store.setFilter('past', {pastEvents: true});
+
+					return [forth, past];
+				}				
 		}],
 		customFields: {
 			fieldSetDialog: "GO.calendar.CustomFieldSetDialog"